+++ WheelWizard/Services/ModManager.cs
@@ -1,19 +1,16 @@
-﻿using Microsoft.Win32;
+﻿using IniParser;
 using System;
 using System.Collections.ObjectModel;
 using System.ComponentModel;
 using System.IO;
 using System.IO.Compression;
 using System.Linq;
-using System.Runtime.CompilerServices;
 using System.Threading.Tasks;
-using System.Windows;
 using WheelWizard.Helpers;
 using WheelWizard.Models.Settings;
 using WheelWizard.Resources.Languages;
 using WheelWizard.Services.Installation;
-using WheelWizard.Services.Launcher;
-using WheelWizard.WPFViews.Popups.Generic;
+using WheelWizard.Views.Popups.Generic;
 
 namespace WheelWizard.Services;
 
@@ -29,7 +26,7 @@ public class ModManager : INotifyPropertyChanged
         private set
         {
             _mods = value;
-            OnPropertyChanged();
+            OnPropertyChanged(nameof(Mods));
         }
     }
 
@@ -40,44 +37,38 @@ public class ModManager : INotifyPropertyChanged
         _mods = new ObservableCollection<Mod>();
     }
 
-    // Events to communicate with the frontend
-    public event Action? ModsLoaded;
-    public event Action? ModsChanged;
-    public event Action? ModProcessingStarted;
-    public event Action? ModProcessingCompleted;
-    public event Action<int, int, string>? ModProcessingProgress;
-    public event Action<string>? ErrorOccurred;
-
     public bool IsModInstalled(int modID)
     {
         return Mods.Any(mod => mod.ModID == modID);
     }
 
-    public async void InitializeAsync()
-    {
-        await LoadModsAsync();
-        ModsLoaded?.Invoke();
-    }
+    public async void ReloadAsync() => await LoadModsAsync();
 
-    public async Task LoadModsAsync()
+    private async Task LoadModsAsync()
     {
         try
         {
-            Mods = await ModInstallation.LoadModsAsync();
+            // Unsubscribe all the old mods and resubscribe all the new mods
             foreach (var mod in Mods)
+            {
+                mod.PropertyChanged -= Mod_PropertyChanged;
+            }
+
+            var newMods = await ModInstallation.LoadModsAsync();
+            foreach (var mod in newMods)
             {
                 mod.PropertyChanged += Mod_PropertyChanged;
             }
-            OnPropertyChanged(nameof(Mods));
-            ModsChanged?.Invoke();
+
+            Mods = newMods;
         }
         catch (Exception ex)
         {
-            ErrorOccurred?.Invoke($"Failed to load mods: {ex.Message}");
+            ErrorOccurred($"Failed to load mods: {ex.Message}");
         }
     }
 
-    public async Task SaveModsAsync()
+    public async void SaveModsAsync()
     {
         try
         {
@@ -85,87 +76,72 @@ public class ModManager : INotifyPropertyChanged
         }
         catch (Exception ex)
         {
-            ErrorOccurred?.Invoke($"Failed to save mods: {ex.Message}");
+            ErrorOccurred($"Failed to save mods: {ex.Message}");
         }
     }
 
     public void AddMod(Mod mod)
     {
-        if (!ModInstallation.ModExists(Mods, mod.Title))
-        {
-            mod.PropertyChanged += Mod_PropertyChanged;
-            Mods.Add(mod);
-            SortModsByPriority();
-            _ = SaveModsAsync(); //this calls async method without awaiting it and no warning :)
-            ModsChanged?.Invoke();
-        }
+        if (ModInstallation.ModExists(Mods, mod.Title)) return;
+
+        mod.PropertyChanged += Mod_PropertyChanged;
+        Mods.Add(mod);
+        SortModsByPriority();
+        SaveModsAsync(); 
     }
 
     public void RemoveMod(Mod mod)
     {
-        if (Mods.Contains(mod))
-        {
-            Mods.Remove(mod);
-            _ = SaveModsAsync();
-            ModsChanged?.Invoke();
-        }
+        if (!Mods.Contains(mod)) return;
+
+        Mods.Remove(mod);
+        SaveModsAsync();
+        OnPropertyChanged(nameof(Mods));
     }
 
     private void Mod_PropertyChanged(object sender, PropertyChangedEventArgs e)
     {
-        if (e.PropertyName == nameof(Mod.IsEnabled) ||
-            e.PropertyName == nameof(Mod.Title) ||
-            e.PropertyName == nameof(Mod.Author) ||
-            e.PropertyName == nameof(Mod.ModID) ||
-            e.PropertyName == nameof(Mod.Priority))
-        {
-            _ = SaveModsAsync();
-            // SortModsByPriority();
-            ModsChanged?.Invoke();
-        }
+        if (e.PropertyName != nameof(Mod.IsEnabled) &&
+            e.PropertyName != nameof(Mod.Title) &&
+            e.PropertyName != nameof(Mod.Author) &&
+            e.PropertyName != nameof(Mod.ModID) &&
+            e.PropertyName != nameof(Mod.Priority)) return;
+        
+        SaveModsAsync();
+        SortModsByPriority();
     }
 
     private void SortModsByPriority()
     {
         var sortedMods = new ObservableCollection<Mod>(Mods.OrderBy(m => m.Priority));
         Mods = sortedMods;
-        OnPropertyChanged(nameof(Mods));
     }
 
     public async void ImportMods()
     {
-        var joinedExtensions = string.Join(";", ModsLaunchHelper.AcceptedModExtensions);
-        joinedExtensions += ";*.zip";
-        var openFileDialog = new OpenFileDialog
-        {
-            Filter = $"Mod files ({joinedExtensions})|{joinedExtensions}|All files (*.*)|*.*",
-            Title = "Select Mod File",
-            Multiselect = true
-        };
+        var fileType = CustomFilePickerFileType.Mods;
+        var selectedFiles = await FilePickerHelper.OpenFilePickerAsync(fileType, allowMultiple: true, title: "Select Mod File");
 
-        if (openFileDialog.ShowDialog() != true) return;
-        var selectedFiles = openFileDialog.FileNames;
-        await ProcessModFilesAsync(selectedFiles);
+        if (selectedFiles.Count == 0) return;
 
+        await ProcessModFilesAsync(selectedFiles.ToArray());
     }
 
     private async Task ProcessModFilesAsync(string[] filePaths)
     {
-        ShowProcessing(true);
         try
         {
             await CombineFilesIntoSingleModAsync(filePaths);
         }
         catch (Exception ex)
         {
-            ErrorOccurred?.Invoke($"Failed to process mod files: {ex.Message}");
+            ErrorOccurred($"Failed to process mod files: {ex.Message}");
         }
-        ShowProcessing(false);
     }
 
     private async Task CombineFilesIntoSingleModAsync(string[] filePaths)
     {
-        var modName = new TextInputPopup("Enter Mod Name").ShowDialog();
+        var modName = await new TextInputWindow().setLabelText("Enter Mod Name").ShowDialog();
         if (!IsValidName(modName)) return;
         var tempZipPath = Path.Combine(Path.GetTempPath(), $"{modName}.zip");
         try
@@ -181,11 +157,10 @@ public class ModManager : INotifyPropertyChanged
             await ModInstallation.InstallModFromFileAsync(tempZipPath, modName, author: "-1", modID: -1);
             if (File.Exists(tempZipPath))
                 File.Delete(tempZipPath);
-            ModProcessingCompleted?.Invoke();
         }
         catch (Exception ex)
         {
-            ErrorOccurred?.Invoke($"Failed to combine and install mod: {ex.Message}");
+            ErrorOccurred($"Failed to combine and install mod: {ex.Message}");
         }
     }
 
@@ -196,53 +171,75 @@ public class ModManager : INotifyPropertyChanged
             mod.IsEnabled = enable;
         }
         _isProcessing = !_isProcessing;
-        ModsChanged?.Invoke();
+        OnPropertyChanged(nameof(Mods));
     }
-
-    public void RenameMod(Mod selectedMod)
-    {
-        var newTitle = new TextInputPopup("Enter Mod Title").ShowDialog();
-        if (!IsValidName(newTitle)) return;
-
-        var oldTitle = selectedMod.Title;
-        var oldDirectoryName = ModInstallation.GetModDirectoryPath(oldTitle);
-        var newDirectoryName = ModInstallation.GetModDirectoryPath(newTitle);
-
-        // Construct the .ini paths
-        var oldIniPath = Path.Combine(oldDirectoryName, $"{oldTitle}.ini"); // Path before directory rename
     
-        try
-        {
-            Directory.Move(oldDirectoryName, newDirectoryName);
-            var updatedOldIniPath = Path.Combine(newDirectoryName, $"{oldTitle}.ini"); // Path after directory rename
-            var newIniPath = Path.Combine(newDirectoryName, $"{newTitle}.ini");
-            if (File.Exists(updatedOldIniPath))
-            {
-                File.Move(updatedOldIniPath, newIniPath);
-            }
+    public async void RenameMod(Mod selectedMod)
+    {
+    var newTitle = await new TextInputWindow().setLabelText("Enter Mod Title").ShowDialog();
 
-            selectedMod.Title = newTitle;
-            SaveModsAsync(); // Ideally: await SaveModsAsync();
-            ModsChanged?.Invoke();
-        }
-        catch (IOException ex)
+    if (!IsValidName(newTitle)) return;
+    
+    var oldTitle = selectedMod.Title;
+
+    var oldDirectoryName = ModInstallation.GetModDirectoryPath(oldTitle);
+    var newDirectoryName = ModInstallation.GetModDirectoryPath(newTitle);
+
+    // Check if the old directory exists
+    if (!Directory.Exists(oldDirectoryName)) return;
+    
+    // var oldIniPath = Path.Combine(oldDirectoryName, $"{oldTitle}.ini");
+    
+    GC.Collect();
+    GC.WaitForPendingFinalizers();
+
+    // Rename the mod directory first
+    try
+    {
+        Directory.Move(oldDirectoryName, newDirectoryName);
+        var newIniPath = Path.Combine(newDirectoryName, $"{newTitle}.ini");
+        var updatedOldIniPath = Path.Combine(newDirectoryName, $"{oldTitle}.ini");
+        if (File.Exists(updatedOldIniPath))
         {
-            ErrorOccurred?.Invoke($"Failed to rename mod directory: {ex.Message}");
+            File.Move(updatedOldIniPath, newIniPath);
+            // Update the mod name inside the .ini file
+            var parser = new FileIniDataParser();
+            var data = parser.ReadFile(newIniPath);
+            data["Mod"]["Name"] = newTitle;
+            parser.WriteFile(newIniPath, data);
+        }
+
+        // Update the selectedMod's title
+        selectedMod.Title = newTitle;
+        await selectedMod.SaveToIniAsync(newIniPath);
+        SaveModsAsync();
+        OnPropertyChanged(nameof(Mods));
+    }
+    catch (IOException ex)
+    {
+        ErrorOccurred($"Failed to rename mod directory: {ex.Message}");
+    }
+    finally
+    {
+        if (Directory.Exists(oldDirectoryName))
+        {
+            Directory.Delete(oldDirectoryName, true);
         }
     }
+ 
+    ReloadAsync();
+}
 
 
-    public void DeleteMod(Mod selectedMod)
+    public async void DeleteMod(Mod selectedMod)
     {
-        var areTheySure = new YesNoWindow().SetMainText(Humanizer.ReplaceDynamic(Phrases.PopupText_SureDeleteQuestion, selectedMod.Title)).AwaitAnswer();
+        var areTheySure = await new YesNoWindow().SetMainText(Humanizer.ReplaceDynamic(Phrases.PopupText_SureDeleteQuestion, selectedMod.Title)).AwaitAnswer();
         if (!areTheySure) return;
 
         var modDirectory = ModInstallation.GetModDirectoryPath(selectedMod.Title);
-        Console.WriteLine($"Attempting to delete mod directory: {modDirectory}");
 
         if (!Directory.Exists(modDirectory))
         {
-            Console.WriteLine($"Mod directory not found: {modDirectory}");
             RemoveMod(selectedMod);
             return;
         }
@@ -261,7 +258,7 @@ public class ModManager : INotifyPropertyChanged
         }
         catch (Exception ex) // Catch a more general exception
         {
-            ErrorOccurred?.Invoke($"Failed to delete mod directory: {ex.Message}");
+            ErrorOccurred($"Failed to delete mod directory: {ex.Message}");
         }
     }
 
@@ -279,7 +276,7 @@ public class ModManager : INotifyPropertyChanged
         }
         else
         {
-            ErrorOccurred?.Invoke(Phrases.PopupText_NoModFolder);
+            ErrorOccurred(Phrases.PopupText_NoModFolder);
         }
     }
     
@@ -289,14 +286,13 @@ public class ModManager : INotifyPropertyChanged
     
         if (modToDelete == null)
         {
-            ErrorOccurred?.Invoke($"No mod found with ID: {modId}");
+            ErrorOccurred($"No mod found with ID: {modId}");
             return;
         }
         DeleteMod(modToDelete);
     }
 
-
-    public void ReorderMod(Mod movedMod, int newIndex)
+    public void MoveMod(Mod movedMod, int newIndex)
     {
         var oldIndex = Mods.IndexOf(movedMod);
         if (oldIndex < 0 || newIndex < 0 || newIndex >= Mods.Count) return;
@@ -307,38 +303,111 @@ public class ModManager : INotifyPropertyChanged
             Mods[i].Priority = i;
         }
         SaveModsAsync();
-        ModsChanged?.Invoke();
+        OnPropertyChanged(nameof(Mods));
     }
 
-    // Helper Methods
 
-    private bool IsValidName(string name)
+    private bool IsValidName(string? name)
     {
         if (string.IsNullOrWhiteSpace(name))
         {
-            ErrorOccurred?.Invoke(Phrases.PopupText_ModNameEmpty);
+            ErrorOccurred(Phrases.PopupText_ModNameEmpty);
             return false;
         }
 
         if (!ModInstallation.ModExists(Mods, name)) return true;
 
-        ErrorOccurred?.Invoke(Humanizer.ReplaceDynamic(Phrases.PopupText_ModNameExists, name));
+        ErrorOccurred(Humanizer.ReplaceDynamic(Phrases.PopupText_ModNameExists, name));
         return false;
     }
 
-    private void ShowProcessing(bool isProcessing)
-    {
-        if (isProcessing)
-            ModProcessingStarted?.Invoke();
-        else
-            ModProcessingCompleted?.Invoke();
-    }
 
+    private void ErrorOccurred(string? errorMessage)
+    {
+        new MessageBoxWindow()
+            .SetMessageType(MessageBoxWindow.MessageType.Error)
+            .SetTitleText("An error occured")
+            .SetInfoText(errorMessage)
+            .Show();
+    }
+    
+    #region PropertyChanged
     public event PropertyChangedEventHandler? PropertyChanged;
 
-    protected virtual void OnPropertyChanged([CallerMemberName] string? propertyName = null)
+    protected virtual void OnPropertyChanged(string propertyName)
     {
         PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
     }
-}
+    #endregion
 
+    public void DecreasePriority(Mod mod)
+    {
+        if (Mods.IndexOf(mod) == -1)
+        {
+            new MessageBoxWindow()
+                .SetMessageType(MessageBoxWindow.MessageType.Error)
+                .SetTitleText("Cannot find mod")
+                .SetInfoText("Cannot find the mod to decrease priority.")
+                .Show();
+            return;
+        }
+        if (mod.Priority == GetLowestActivePriority() || Mods.Count == 1)
+        {
+            new MessageBoxWindow()
+                .SetMessageType(MessageBoxWindow.MessageType.Warning)
+                .SetTitleText("Cannot decrease priority")
+                .SetInfoText("Cannot decrease priority of the first mod.")
+                .Show();
+            return;
+        }
+ 
+        // Find mod with next lower priority value
+        var modAbove = Mods.Where(m => m.Priority < mod.Priority)
+            .OrderByDescending(m => m.Priority)
+            .FirstOrDefault();
+        if (modAbove == null) return;
+        
+        (modAbove.Priority, mod.Priority) = (mod.Priority, modAbove.Priority);
+
+        SortModsByPriority();
+        SaveModsAsync();
+    }
+
+    public void IncreasePriority(Mod mod)
+    {
+        if (Mods.IndexOf(mod) == -1)
+        {
+            new MessageBoxWindow()
+                .SetMessageType(MessageBoxWindow.MessageType.Error)
+                .SetTitleText("Cannot find mod")
+                .Show();
+            return;
+        }
+    
+        if (mod.Priority == GetHighestActivePriority() || Mods.Count == 1)
+        {
+            new MessageBoxWindow()
+                .SetMessageType(MessageBoxWindow.MessageType.Warning)
+                .SetTitleText("Cannot increase priority")
+                .SetInfoText("Cannot increase priority of the last mod.")
+                .Show();
+            return;
+        }
+    
+        // Find mod with next higher priority value
+        var modBelow = Mods.Where(m => m.Priority > mod.Priority)
+            .OrderBy(m => m.Priority)
+            .FirstOrDefault();
+    
+        if (modBelow == null) return; // Should not happen but just in case
+    
+        // Swap priorities
+        (modBelow.Priority, mod.Priority) = (mod.Priority, modBelow.Priority);
+
+        SortModsByPriority();
+        SaveModsAsync();
+    }
+    
+    public int GetLowestActivePriority() =>Mods.Min(m => m.Priority);
+    public int GetHighestActivePriority() => Mods.Max(m => m.Priority);
+}
